chore: rename unsee scripts
fix: album name identification
