+++ main.py
@@ -5,9 +5,9 @@ import sys
 from typing import List
 
 from unsee_dl import __version__ as unsee_dl_version
-from unsee_dl.unsee import get_album_id_from_url, is_beta_album_id
-from unsee_dl.unsee_beta import Client as ClientBeta
-from unsee_dl.unsee_dl import Client
+from unsee_dl.unsee import get_album_id_from_url, is_old_album_id
+from unsee_dl.unsee_new import Client as ClientNew
+from unsee_dl.unsee_old import ClientOld
 
 
 def main():
@@ -15,7 +15,7 @@ def main():
 
 
 async def download_old(album_ids: List[str], out_dir: str, group_album: bool):
-    async with Client(out_path=out_dir, group_album=group_album) as client:
+    async with ClientOld(out_path=out_dir, group_album=group_album) as client:
         for album_id in album_ids:
             # noinspection PyBroadException
             try:
@@ -28,8 +28,8 @@ async def download_old(album_ids: List[str], out_dir: str, group_album: bool):
                 )
 
 
-async def download_beta(album_ids: List[str], out_dir: str, group_album: bool):
-    async with ClientBeta(out_path=out_dir, group_album=group_album) as client:
+async def download_new(album_ids: List[str], out_dir: str, group_album: bool):
+    async with ClientNew(out_path=out_dir, group_album=group_album) as client:
         await client.anonymous_login()
 
         for album_id in album_ids:
@@ -88,17 +88,18 @@ async def run_downloader():
 
     # Download images
     album_ids = [get_album_id_from_url(url) for url in args.album_ids]
-    album_ids_old_version = list(filter(lambda x: not is_beta_album_id(x), album_ids))
-    album_ids_beta_version = list(filter(lambda x: is_beta_album_id(x), album_ids))
+    album_ids_old_version = list(filter(lambda x: is_old_album_id(x), album_ids))
+    album_ids_new_version = list(filter(lambda x: not is_old_album_id(x), album_ids))
 
     if len(album_ids_old_version) > 0:
         await download_old(album_ids_old_version, args.out_dir, args.group_album)
 
-    if len(album_ids_beta_version) > 0:
-        await download_beta(album_ids_beta_version, args.out_dir, args.group_album)
-
-    logging.shutdown()
+    if len(album_ids_new_version) > 0:
+        await download_new(album_ids_new_version, args.out_dir, args.group_album)
 
 
 if __name__ == "__main__":
-    main()
+    try:
+        main()
+    finally:
+        logging.shutdown()
